Fixed warnings in windows compilation [HZ-5297] (#1385)
Fixed warning in windows compilation. Most warnings were due to type
conversions and dll imports.

fixes #1350

---------

+++ hazelcast/src/hazelcast/client/metrics.cpp
@@ -103,9 +103,11 @@ zlib_compress(const std::vector<byte>& input)
         output.push_back(
           static_cast<byte>(is_final));         // BFINAL = is_final, BTYPE = 00
         output.push_back(block_size & 0xff);    // LEN - least significant
-        output.push_back(block_size >> 8);      // LEN - most significant
+        output.push_back(
+          static_cast<byte>(block_size >> 8));  // LEN - most significant
         output.push_back((~block_size) & 0xff); // NLEN - least significant
-        output.push_back((~block_size) >> 8);   // NLEN - most significant
+        output.push_back(
+          static_cast<byte>((~block_size) >> 8)); // NLEN - most significant
 
         // copy uncompressed bytes and accumulate checksum
         for (std::size_t i = block_start; i < block_end; i++) {
@@ -126,9 +128,9 @@ zlib_compress(const std::vector<byte>& input)
     }
 
     // Adler32 checksum
-    output.push_back(a2 >> 8);
+    output.push_back(static_cast<byte>(a2 >> 8));
     output.push_back(a2 & 0xff);
-    output.push_back(a1 >> 8);
+    output.push_back(static_cast<byte>(a1 >> 8));
     output.push_back(a1 & 0xff);
 
     return output;
